document \[...] form for order_by
I am not sure if this is the best way to note that this works, but I had
assumed that this list was fairly exhaustive until I tried \[...] and
found it worked.  This is just one way to indicate that it will work.
Any such note may help future users, though.