modify belch and puke to operate as method calls
SQL::Abstract::Classic made this change and honestly I'm not sure why we
didn't do that sooner.

Either way, though, (a) current DBIx::Class depends on that change so would
fail against mainline SQL::Abstract as a result (b) it's a good idea so the
spirit of open source demands stealing it.